specs/001-llm-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pydantic==2.5.0
pyyaml==6.0.1
click==8.1.7
python-multipart==0.0.6
requests==2.31.0